moved XSLT setting to CityGML export options in config
