+++ impexp-client/src/main/java/org/citydb/gui/modules/common/XSLTransformationPanel.java
@@ -73,7 +73,8 @@ public class XSLTransformationPanel extends AbstractPreferencesComponent {
 
     @Override
     public boolean isModified() {
-        XSLTransformation transformation = isExport ? config.getExportConfig().getXSLTransformation() :
+        XSLTransformation transformation = isExport ?
+                config.getExportConfig().getCityGMLOptions().getXSLTransformation() :
                 config.getImportConfig().getCityGMLOptions().getXSLTransformation();
 
         if (transformation.isEnabled() != applyStylesheets.isSelected()) return true;
@@ -114,7 +115,7 @@ public class XSLTransformationPanel extends AbstractPreferencesComponent {
     @Override
     public void setSettings() {
         XSLTransformation transformation = isExport ?
-                config.getExportConfig().getXSLTransformation() :
+                config.getExportConfig().getCityGMLOptions().getXSLTransformation() :
                 config.getImportConfig().getCityGMLOptions().getXSLTransformation();
 
         transformation.setEnabled(applyStylesheets.isSelected());
@@ -158,7 +159,7 @@ public class XSLTransformationPanel extends AbstractPreferencesComponent {
     @Override
     public void loadSettings() {
         XSLTransformation transformation = isExport ?
-                config.getExportConfig().getXSLTransformation() :
+                config.getExportConfig().getCityGMLOptions().getXSLTransformation() :
                 config.getImportConfig().getCityGMLOptions().getXSLTransformation();
 
         applyStylesheets.setSelected(transformation.isEnabled());
